Explicitly specify project root for poetry
Projects with nested `pyproject.toml` may pick up the incorrect file
relative to the `.mise(.local)?.toml` if the working directory is deeper
than a nested file.

This can lead to a (shared) virtual-environment at the top-level not
being activated.

```
project/         <-- VIRTUAL_ENV/POETRY_ACTIVE set as expected
  .mise.toml
  .venv/         <-- shared by sub-projects
  pyproject.toml
  sub-project/   <-- VIRTUAL_ENV/POETRY_ACTIVE not set
    pyproject.toml
```

This change infers the poetry project directory from the
`pyproject.toml` file. If virtual-environments are desired for the
sub-projects, then a `.mise.toml` can be created in those directories.